refactor: オフラインユーザアイテムも別コンポーネント化
